--- u-music/.env
@@ -1,2 +1,0 @@
-REACT_APP_SPOTIFY_CLIENT_ID=78cf2caf0c014010ba9267597eaac6a3
-REACT_APP_SPOTIFY_CLIENT_SECRET=660ffced675f4b5bb92b8f8312dcdeb5
